+++ resolver/CHANGELOG.md
@@ -19,7 +19,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - ipv4_lookup for looking up *only* ipv4 (lookup_ip has options for dual-stack)
 - ipv6_lookup for looking up *only* ipv6 (lookup_ip has options for dual-stack)
 - mx_lookup for querying mail exchanges
-- srv_lookup for service records
+- srv_lookup for service records and also a specialized form for ease of use lookup_service
 - txt_lookup for text record lookups
 
 ## 0.4.0